requirements.txt
certifi==2019.11.28
dnspython==1.16.0
mysql-connector-python==8.0.19
numpy==1.18.1
pandas==1.0.0
pipenv==2018.11.26
protobuf==3.6.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
